Move platform version os modules to their own files
